KAF002: add version check

+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.infra/src/main/java/nts/uk/ctx/at/request/infra/repository/application/common/JpaApplicationRepository.java
@@ -68,7 +68,7 @@ public class JpaApplicationRepository extends JpaRepository implements Applicati
 				new KafdtApplicationPK(
 					domain.getCompanyID(), 
 					domain.getApplicationID()), 
-				(int)(long)domain.getVersion(),
+				domain.getVersion(),
 				appReasonID,
 				domain.getPrePostAtr().value, 
 				domain.getInputDate() ,
